+++ tests/Unit/Api/IssuePriority/ListTest.php
@@ -5,6 +5,7 @@ namespace Redmine\Tests\Unit\Api\IssuePriority;
 use PHPUnit\Framework\TestCase;
 use Redmine\Api\IssuePriority;
 use Redmine\Client\Client;
+use Redmine\Exception\UnexpectedResponseException;
 
 /**
  * @covers \Redmine\Api\IssuePriority::list
@@ -62,5 +63,30 @@ class ListTest extends TestCase
 
         // Perform the tests
         $this->assertSame($expectedReturn, $api->list($allParameters));
+    }
+
+    public function testListThrowsException()
+    {
+        // Create the used mock objects
+        $client = $this->createMock(Client::class);
+        $client->expects($this->exactly(1))
+            ->method('requestGet')
+            ->with('/enumerations/issue_priorities.json')
+            ->willReturn(true);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseBody')
+            ->willReturn('');
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseContentType')
+            ->willReturn('application/json');
+
+        // Create the object under test
+        $api = new IssuePriority($client);
+
+        $this->expectException(UnexpectedResponseException::class);
+        $this->expectExceptionMessage('The Redmine server responded with an unexpected body.');
+
+        // Perform the tests
+        $api->list();
     }
 }